Resolves: #10178 

<!-- Use "x" to fill the checkboxes below like [x] -->

- [x] I signed [CLA](https://www.audacityteam.org/cla/)
- [x] The title of the pull request describes an issue it addresses
- [x] If changes are extensive, then there is a sequence of easily
reviewable commits
- [x] Each commit's message describes its purpose and effects
- [x] There are no behavior changes unnecessary for the stated purpose
of the PR

Recommended:
- [x] Each commit compiles and runs on my machine without known
undesirable changes of behavior

QA:

- [ ] Autobot test cases have been run

+++ src/spectrogram/internal/au3/au3trackspectrogramconfiguration.cpp
@@ -32,22 +32,22 @@ Au3TrackSpectrogramConfiguration::Au3TrackSpectrogramConfiguration(Au3Spectrogra
     : m_settings(settings)
 {}
 
-int Au3TrackSpectrogramConfiguration::minFreq() const
+double Au3TrackSpectrogramConfiguration::minFreq() const
 {
     return m_settings.minFreq;
 }
 
-void Au3TrackSpectrogramConfiguration::setMinFreq(int value)
+void Au3TrackSpectrogramConfiguration::setMinFreq(double value)
 {
     m_settings.minFreq = value;
 }
 
-int Au3TrackSpectrogramConfiguration::maxFreq() const
+double Au3TrackSpectrogramConfiguration::maxFreq() const
 {
     return m_settings.maxFreq;
 }
 
-void Au3TrackSpectrogramConfiguration::setMaxFreq(int value)
+void Au3TrackSpectrogramConfiguration::setMaxFreq(double value)
 {
     m_settings.maxFreq = value;
 }
