+++ src/spectrogram/internal/au3/au3trackspectrogramconfiguration.h
@@ -20,11 +20,11 @@ public:
     Au3TrackSpectrogramConfiguration(Au3SpectrogramSettings&);
     ~Au3TrackSpectrogramConfiguration() override = default;
 
-    int minFreq() const override;
-    void setMinFreq(int value) override;
+    double minFreq() const override;
+    void setMinFreq(double value) override;
 
-    int maxFreq() const override;
-    void setMaxFreq(int value) override;
+    double maxFreq() const override;
+    void setMaxFreq(double value) override;
 
     int colorGainDb() const override;
     void setColorGainDb(int value) override;
